#3209 Another WI, refactoring done
* Remaining problem: splitting on top level CUs or on a singular CU does
not instantly refresh
* Problem 2: I am not sure the console floods problem was solved
https://github.com/metasfresh/metasfresh/issues/3209
